Move tests from `ConstraintTest` to `AttributeTest`
